fix: exception handling, edit replyrequest, authorization processing

+++ src/main/java/com/example/Mini_Project1/service/CommentService.java
@@ -4,6 +4,10 @@ import com.example.Mini_Project1.entity.Comment;
 import com.example.Mini_Project1.entity.Course;
 import com.example.Mini_Project1.entity.User;
 import com.example.Mini_Project1.entity.Payment;
+import com.example.Mini_Project1.exception.AuthenticationException;
+import com.example.Mini_Project1.exception.BadRequestException;
+import com.example.Mini_Project1.exception.PaymentRequiredException;
+import com.example.Mini_Project1.exception.ResourceNotFoundException;
 import com.example.Mini_Project1.repository.CommentRepository;
 import com.example.Mini_Project1.repository.CourseRepository;
 import com.example.Mini_Project1.repository.UserRepository;
@@ -38,30 +42,28 @@ public class CommentService {
     public CommentResponse createComment(CommentRequest request) {
         // Kiểm tra nếu content là null
         if (request.getContent() == null) {
-            throw new IllegalArgumentException("Content cannot be null");
+            throw new BadRequestException("Content cannot be null");
         }
 
         UUID courseId = null;
         try {
             courseId = UUID.fromString(String.valueOf(request.getCourseId()));
         } catch (IllegalArgumentException e) {
-            throw new IllegalArgumentException("Invalid UUID format for CourseId");
+            throw new BadRequestException("Invalid UUID format for CourseId");
         }
 
-        // Lấy user từ
         UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        User user = userRepository.findById(userDetails.getUsername())  // userDetails.getUsername() là userId trong JWT
-                .orElseThrow(() -> new RuntimeException("User not found"));
+        User user = userRepository.findById(userDetails.getUsername())
+                .orElseThrow(() -> new ResourceNotFoundException("User not found"));
 
         // Kiểm tra sự tồn tại của userId và courseId trong bảng Payment
         Payment payment = paymentRepository.findByUserIdAndCourseId(user.getId(), courseId.toString());
         if (payment == null) {
-            throw new RuntimeException("User is not authorized to comment. Please make sure payment exists for this course.");
+            throw new PaymentRequiredException("User is not authorized to comment. Please make sure payment exists for this course.");
         }
 
-        // Tìm khóa học từ cơ sở dữ liệu
         Course course = courseRepository.findById(courseId.toString())
-                .orElseThrow(() -> new RuntimeException("Course not found"));
+                .orElseThrow(() -> new PaymentRequiredException("Course not found"));
 
         // Tạo mới comment
         Comment newComment = Comment.builder()
@@ -72,26 +74,22 @@ public class CommentService {
                 .createdDate(new Date())
                 .updatedDate(new Date())
                 .build();
-
-        // Lưu comment vào cơ sở dữ liệu
         commentRepository.save(newComment);
 
-        // Trả về CommentResponse
         return modelMapper.map(newComment, CommentResponse.class);
     }
     public CommentResponse replyToComment(ReplyRequest request, UserDetails userDetails) {
-        // Tìm comment cha từ cơ sở dữ liệu
         Comment parentComment = commentRepository.findById(request.getRootCommentId())
-                .orElseThrow(() -> new RuntimeException("Parent comment not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("Parent comment not found"));
 
         // Lấy user từ UserDetails
         User user = userRepository.findById(userDetails.getUsername())  // userDetails.getUsername() là userId
-                .orElseThrow(() -> new RuntimeException("User not found"));
-
+                .orElseThrow(() -> new ResourceNotFoundException("User not found"));
+        String courseId = parentComment.getCourse().getId();
         // Kiểm tra xem người dùng có thanh toán cho khóa học này không
         Payment payment = paymentRepository.findByUserIdAndCourseId(user.getId(), parentComment.getCourse().getId());
         if (payment == null) {
-            throw new RuntimeException("User has not completed payment for this course");
+            throw new PaymentRequiredException("User has not completed payment for this course");
         }
 
         // Tạo mới comment trả lời
@@ -112,18 +110,38 @@ public class CommentService {
 
 
     // Lấy danh sách tất cả comment
-    public List<CommentResponse> getAllComments() {
-        List<Comment> comments = commentRepository.findAll();
-        return comments.stream()
+    public List<CommentResponse> getAllComments(UserDetails userDetails) {
+        User user = userRepository.findById(userDetails.getUsername())
+                .orElseThrow(() -> new ResourceNotFoundException("User not found"));
+
+        List<Comment> allComments = commentRepository.findAll();
+        List<CommentResponse> commentResponses = allComments.stream()
+                .filter(comment -> {
+                    Course course = comment.getCourse();
+                    if (course == null) {
+                        return false;
+                    }
+
+                    Payment payment = paymentRepository.findByUserIdAndCourseId(user.getId(), course.getId());
+                    return payment != null;  // Chỉ lọc các bình luận thuộc khóa học mà người dùng đã thanh toán
+                })
                 .map(comment -> modelMapper.map(comment, CommentResponse.class))
                 .collect(Collectors.toList());
+
+        return commentResponses;
     }
 
     // Cập nhật nội dung comment
-    public CommentResponse updateCommentContent(UUID commentId, UpdateCommentRequest request) {
+    public CommentResponse updateCommentContent(UUID commentId, UpdateCommentRequest request, UserDetails userDetails) {
         Comment comment = commentRepository.findById(commentId.toString())
-                .orElseThrow(() -> new RuntimeException("Comment not found"));
-
+                .orElseThrow(() -> new ResourceNotFoundException("Comment not found"));
+        if (!comment.getUser().getId().equals(userDetails.getUsername())) {
+            throw new AuthenticationException("You are not authorized to update this comment.");
+        }
+        Payment payment = paymentRepository.findByUserIdAndCourseId(userDetails.getUsername(), comment.getCourse().getId());
+        if (payment == null) {
+            throw new PaymentRequiredException("You must complete the payment for this course to update the comment.");
+        }
         // Cập nhật nội dung comment và thời gian sửa đổi
         comment.setContent(request.getContent());
         comment.setUpdatedDate(new Date());
@@ -133,40 +151,53 @@ public class CommentService {
     }
 
     // Xóa comment
+    // Xóa comment
     @Transactional
-    public String deleteComment(UUID commentId) {
+    public String deleteComment(UUID commentId, UserDetails userDetails) {
+        // Lấy comment cần xóa
         Comment comment = commentRepository.findById(commentId.toString())
-                .orElseThrow(() -> new RuntimeException("Comment not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("Comment not found"));
 
+
+        if (!comment.getUser().getId().equals(userDetails.getUsername()))  {
+            throw new AuthenticationException("You are not authorized to delete this comment.");
+        }
         deleteReplies(commentId.toString());
-
         commentRepository.delete(comment);
 
         return "Comment and its replies deleted successfully";
     }
 
+    // Phương thức phụ trợ để xóa các reply
     private void deleteReplies(String rootCommentId) {
         List<Comment> replies = commentRepository.findByRootComment(rootCommentId);
         for (Comment reply : replies) {
+            // Đệ quy để xóa các reply của reply
             deleteReplies(reply.getId());
             commentRepository.delete(reply);
         }
     }
 
-    public List<CommentResponse> getCommentsByCourseId(String courseId) {
-        // Kiểm tra UUID format
+
+
+    public List<CommentResponse> getCommentsByCourseId(String courseId, UserDetails userDetails) {
         UUID courseUUID;
         try {
             courseUUID = UUID.fromString(courseId);
         } catch (IllegalArgumentException e) {
-            throw new IllegalArgumentException("Invalid UUID format for CourseId");
+            throw new BadRequestException("Invalid UUID format for CourseId");
         }
 
-        // Lấy danh sách comment từ repository
+        Payment payment = paymentRepository.findByUserIdAndCourseId(userDetails.getUsername(), courseId);
+        if (payment == null) {
+            throw new PaymentRequiredException("You must complete the payment for this course to view comments.");
+        }
+
+
         List<Comment> comments = commentRepository.findByCourseId(courseUUID.toString());
         return comments.stream()
                 .map(comment -> modelMapper.map(comment, CommentResponse.class))
                 .collect(Collectors.toList());
     }
-
 }
+
